refactor: complete Phase 2 SRP refactoring with window_operations extraction
- Successfully extracted all window state management to window_operations.rs
- Window manager now delegates all operations to appropriate modules
- Reduced window_manager.rs complexity through clean separation of concerns
- All tests pass and clippy warnings resolved
- Total lines reduced from 1000+ to 740 in window_manager.rs
